+++ packages/graphic-interface/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @mlightcad/graphic-interface
 
+## 3.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @mlightcad/geometry-engine@3.0.1
+
 ## 3.0.0
 
 ### Minor Changes